Port original collections consolidation
